+++ example/lib/src/app/view.dart
@@ -4,6 +4,7 @@ export 'package:mvc_application_example/src/app/view/app.dart';
 // application's menu bar
 export 'package:mvc_application_example/src/app/view/menu/app_menu.dart';
 
+// Language Spinner
 export 'package:mvc_application_example/src/app/view/menu/iso_spinner.dart';
 
 // Color picker routine